Don't try to use std::exception::what
The problem with what() returning a const char * is that nobody knows what is
supposed to free this pointer. I could add some nasty logic for updating a
mutable, pre-declared member in the what() (or maybe even in the constructor),
but that would be rather ugly. Let's just use a custom function here.