Rename the parser test fixture to a reasonable name
We'll have different fixtures in future, so it makes sense to be able to use
them together (if only for much simplified linking). This patch is the first
setp towards that, giving the F fixture a reasonable name from which it is
possible to actually find out what its goal is.