Rework how we access the streams for JSON IO
This commit changes the way how we access the iostreams for JSON parsing and
sending. The goal here is to be able to access a copy of the JSON data
"ex-post", after the parsing has already failed.  It is not particularly
elegant, but at least it returns the offending JSON data. I believe this is very
much worth the effort.

Fixes #202