Wrapper classes for object attributes
When parsing JSON into object attributes, we have to check the individual JSON
data for validity. Given that we have to somehow pass along the type
information, we cannot just use the existing jsonValueToDeskaValue(), but
instead extract into a wrapper class for each individual attribute which
performs a type-check, and then copy the result back.

refs #203 @50min