Force bash as the interpreter of the run-testcase wrapper
It turned out that it's probably a shell's job to handle shebangs, and one can't
blame cmake for not really fully supporting them. Let's make sure we always ask
bash to run the wrapper, even though its shebang was right.

Thanks to Tomas for reporting the issue and to his system's developers to make
/bin/sh different from bash :).