Fix build failure on RHEL5
Looks like I have to make that function inline in order for it to be visible
among multiple translation units. Also had to restore extraction of boost's
ptime, which I removed by accident in f702845750fece5552808f225d769391badce093.
I'm not sure how come that it succeeded building and the unit tests, maybe I was
just lucky with some default implementation.

Thanks to Tomas for reporting a build failure on his box.