Allow skipping the DB init altogether via DESKA_SKIP_DB_INIT
Please note that this has a potential to be very intrusive operation which will
cause many test cases to fail, as it preserves the database state across the
test runs.  Therefore, any state which is not explicitly ready for *any* state
of the DB could fail randomly.

To use this approach, simply run:
    `DESKA_SKIP_DB_INIT=1 ctest --output-on-failure`

Make sure that the database is already initialized before you run that, though.
A reasonably safe way to achieve that is simply letting a plain old `ctest`
finish at first. Killing a test case via Ctrl-C is known to cause the DB in an
unusable state.