Add some support for code coverage testing
Simply build with -DCMAKE_BUILD_TYPE=Coverage. It will produce a debug build
which additionally writes out profiling information for the gcov. The ctest is
pretty well integrated, and simply running `make experimental` uploads the
coverage results to the dashboard.

One thing which remains to be solved is the gcov resetting; I believe that the
numbers shall be cleared before each `make test` run, but apparently that does
not happen. This should be fixed, so that no manual removal of **.gcov and
**gcda is required.