Refactor the way the configdiff gets called and processed
Rationale: right now the code for actually showing the diff to the user is
pretty compact and short. In future, it's likely going to be much more complex,
to the point that it is longer than the parameter parsing stuff. Therefore I
believe that it makes sense to have it this way.