Deska::Db::Value shall always be constructed from std::string and *never* from const char[]
This is required for adding a bool to the list of types supported by
Deska::Db::Value.  Unfortunately, the bool constructor gets prefered over the
std::string one when the Deska::Db::Value is constructed from a string literal.
Oh well...

I shall probably mention that this was not particularly funny to debug. Dealing
with backtraces like http://pastebin.com/ZHMPzmFR did not help much.