Spell out the expected data
It turns out that the test results do not conform to the DBAPI spec, and yet the
test succeeds.  This means that it is not a wise choice to mess with some magic
conversion from a command list to the expected values, so let's provide the
expected results by hand.

Yes, it's much more work, but at least it catches the wrong results.