cli: move code around in preparation for a shared library unification
What I want to do is to merge the CLI into a single library. Unfortunately, I
get linker errors from undefined ModificationComparatorLesss::operator()
instances when I do that in an obvious way, which is why I want to move all
users of that template into a single translation unit (enumerating all of the
possibilities by hand is a big no-no).  This is the smallest changeset which
does exactly that and still compiles. Oh well.