installer: remove useless privilege granting
The installer has to run as the postgres superuser due to the security
implications of the PgPython language. This means that there isn't much point in
granting this user additional privileges during installation.

This will have to change when we add support for proper user roles to the
Postgres code (refs #325).