Implement IO tracing and replay for the Deska server
This is intended to aid in the CLI debugging. We start with capturing the
troublesome IO stream (very likely by running the server with the
DESKA_SERVER_IO_TRACE variable set).  Then run the CLI one more time, but
instead of directing it against the real Deska server, target it at the
`deska-fake-server path/to/your/trace`.