fzu: Add a way to specify the model of a box directly, without going through an extra instance
I feel a bit awkward making this change, as the whole idea was to push the
information about the outer dimensions, inner structure etc. of a box into an
extra object.  However, this led to having to instantiate that special object
for each metal enclosure, and that would very soon lead to a huge number of
completely useless objects (one more for each pair of twin nodes, for example).

This hack allows the user to specify the model of a box directly. I've tried to
make sure that the XOR between all of the containing objects and the direct
specification is maintained.