Implement command batching at the DBAPI level
The DeskaDb library theoretically has all the information it requires to
determine whether a particular command can be cached or not.  In reality, this
is not a trivial problem (refs #404)

This changeset cuts the time required to restore the current DB snapshot from
some 1min20s to mere 50s. It is possible to cut this further down to ~35 seconds
by grouping all modifying commands together, and if we wanted to take an extra
step and eliminate all of the useless kindInstances() calls, we'd get even
further, but I haven't done that yet.

refs #382