cmake: disable extra warnings about wrong aliasing
The code works and there's an unhealthy level of uncertainity in various ML
threads and bugreports about the underlying issue. I have no idea whether Boost
is to be blamed here, or if it is just a random gcc mishap. Given that
everything seems to work fine (except for the build warnings), I'm hereby just
muting this warning instead of going the full -fno-strict-aliasing route.