doc: fix table labels
Sorry, my bad. Labels outside of the {longtab} do not belong to the tables while
labels inside their headers cause troubles when the table spans several pages.
The correct fix is apparently to put the label into the table, but outside of
its header, and this place works well.

Thanks to Pavel Martinasek for pointing this out.