doc: conclusion: fix the ORM bits to reflect reality
It turns out that Hibernate offers this thing through the Envers module (which
is now in its core).  However, it still appears that our solution has a few
extra parts (like being able to work on any revision without any extra effort
besides specifying the version number) which deserve the "unique" status.

A huge thanks to Pavel Martinasek for mentioning that.